+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Version 5.6-SNAPSHOT
 
 CHANGED:
 
+- update of jobj-compare dependency to new version 3.4
+- update of silly-collections dependency to new version 8.1
+- update of commons-lang3 dependency to new version 3.10
+- update of test dependency testng to new version 7.2.0
 - update of test dependency meanbean-factories to new version 1.2
 
 Version 5.5